Move data readers to their own module
