+++ backend/server/data_processors/__init__.py
@@ -1,11 +1,8 @@
-"""Module for data readers and transformers"""
+"""Module for data transformers"""
 
-from .betting_data_reader import BettingDataReader
-from .match_data_reader import MatchDataReader
 from .data_cleaner import DataCleaner
 from .team_data_stacker import TeamDataStacker
 from .feature_builder import FeatureBuilder
 from .player_data_stacker import PlayerDataStacker
-from .fitzroy_data_reader import FitzroyDataReader
 from .player_data_aggregator import PlayerDataAggregator
 from .oppo_feature_builder import OppoFeatureBuilder
