+++ backend/server/data_readers/fitzroy_data_reader.py
